requirements.txt
paho-mqtt
pygame
pyopengl
serial
